Add client side component to client status monitor. Clients now support live reload of setting values.

+++ src/api/Fig.Datalayer/BusinessEntities/ClientRunSessionBusinessEntity.cs
@@ -3,15 +3,15 @@ namespace Fig.Datalayer.BusinessEntities;
 public class ClientRunSessionBusinessEntity
 {
     public virtual Guid Id { get; set; }
-    
+
     public virtual Guid RunSessionId { get; set; }
-    
+
     public virtual DateTime? LastSeen { get; set; }
 
     public virtual bool? LiveReload { get; set; }
 
-    public virtual int? PollIntervalMs { get; set; }
-    
+    public virtual double? PollIntervalMs { get; set; }
+
     public virtual double UptimeSeconds { get; set; }
 
     public virtual string? IpAddress { get; set; }
