+++ src/common/Fig.Contracts/Status/ClientRunSessionDataContract.cs
@@ -5,13 +5,13 @@ namespace Fig.Contracts.Status
     public class ClientRunSessionDataContract
     {
         public Guid RunSessionId { get; set; }
-    
+
         public DateTime? LastSeen { get; set; }
 
         public bool? LiveReload { get; set; }
 
-        public int? PollIntervalMs { get; set; }
-    
+        public double? PollIntervalMs { get; set; }
+
         public double UptimeSeconds { get; set; }
 
         public string? IpAddress { get; set; }
